Skip heroku features specs on the CI
We are not able to run the stubed version of the heroku on the Travis CI.
This [1] is the last build that we are able to be green on the CI all new PRs
are getting a build error.

After some reasearch and debug I don't know why in the Travis env we are not
able to run our stub version of the heroku. This [2] is a similar issue.

To fix the problem for now we only run this stub version of the heroku on local
machines until we find a good solution to that.

[1]: https://travis-ci.org/Helabs/pah/builds/102085948
[2]: http://stackoverflow.com/questions/33239151/setup-of-path-when-running-aruba-tests-on-travis-ci